+++ src/nationalarchives/components/card/card.stories.js
@@ -2,24 +2,33 @@ import Card from "./template.njk";
 import macroOptions from "./macro-options.json";
 
 const argTypes = {
-  heading: { control: "object" },
+  title: { control: "text" },
+  supertitle: { control: "text" },
+  headingLevel: { control: { type: "number", min: 1, max: 6 } },
+  headingSize: { control: "inline-radio", options: ["m", "l"] },
   href: { control: "text" },
-  image: { control: "object" },
+  imageSrc: { control: { type: "file", accept: ".jpg" } },
+  imageAlt: { control: "text" },
+  imageWidth: { control: { type: "number", min: 1 } },
+  imageHeight: { control: { type: "number", min: 1 } },
+  imageSources: { control: "object" },
   label: { control: "text" },
   body: { control: "text" },
   text: { control: "text" },
   actions: { control: "object" },
   horizontal: { control: "boolean" },
-  style: { control: "text" },
+  style: { control: "inline-radio", options: ["none", "boxed", "accent"] },
   htmlElement: { control: "text" },
   classes: { control: "text" },
   attributes: { control: "object" },
 };
 
 Object.keys(argTypes).forEach((argType) => {
-  argTypes[argType].description = macroOptions.find(
-    (option) => option.name === argType,
-  )?.description;
+  const macroOption = macroOptions.find((option) => option.name === argType);
+  argTypes[argType] = {
+    ...argTypes[argType],
+    description: macroOption?.description,
+  };
 });
 
 export default {
@@ -28,9 +37,16 @@ export default {
 };
 
 const Template = ({
-  heading,
+  title,
+  supertitle,
+  headingLevel,
+  headingSize,
   href,
-  image,
+  imageSrc,
+  imageAlt,
+  imageWidth,
+  imageHeight,
+  imageSources,
   label,
   body,
   text,
@@ -43,9 +59,16 @@ const Template = ({
 }) =>
   Card({
     params: {
-      heading,
+      title,
+      supertitle,
+      headingLevel,
+      headingSize,
       href,
-      image,
+      imageSrc,
+      imageAlt,
+      imageWidth,
+      imageHeight,
+      imageSources,
       label,
       body,
       text,
@@ -60,47 +83,34 @@ const Template = ({
 
 export const Standard = Template.bind({});
 Standard.args = {
-  heading: {
-    supertitle: "Card supertitle",
-    title: "Card title",
-    level: 3,
-    size: "m",
-  },
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "m",
   href: "#",
-  image: {
-    src: "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
-    alt: "The National Archives office",
-    width: 1996,
-    height: 1331,
-  },
+  imageSrc:
+    "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
+  imageAlt: "The National Archives office",
+  imageWidth: 1996,
+  imageHeight: 1331,
   label: "New",
   body: "<p>Card body</p>",
-  // actions: [
-  //   {
-  //     text: "Card action",
-  //     href: "#",
-  //     title: "Go and do the action",
-  //   },
-  // ],
   htmlElement: "article",
   classes: "tna-card--demo",
 };
 
 export const Boxed = Template.bind({});
 Boxed.args = {
-  heading: {
-    supertitle: "Card supertitle",
-    title: "Card title",
-    level: 3,
-    size: "m",
-  },
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "m",
   href: "#",
-  image: {
-    src: "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
-    alt: "The National Archives office",
-    width: 1996,
-    height: 1331,
-  },
+  imageSrc:
+    "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
+  imageAlt: "The National Archives office",
+  imageWidth: 1996,
+  imageHeight: 1331,
   label: "New",
   body: "<p>Card body</p>",
   style: "boxed",
@@ -110,19 +120,16 @@ Boxed.args = {
 
 export const Accent = Template.bind({});
 Accent.args = {
-  heading: {
-    supertitle: "Card supertitle",
-    title: "Card title",
-    level: 3,
-    size: "m",
-  },
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "m",
   href: "#",
-  image: {
-    src: "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
-    alt: "The National Archives office",
-    width: 1996,
-    height: 1331,
-  },
+  imageSrc:
+    "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
+  imageAlt: "The National Archives office",
+  imageWidth: 1996,
+  imageHeight: 1331,
   label: "New",
   body: "<p>Card body</p>",
   style: "accent",
@@ -132,18 +139,15 @@ Accent.args = {
 
 export const Horizontal = Template.bind({});
 Horizontal.args = {
-  heading: {
-    supertitle: "Card supertitle",
-    title: "Card title",
-    level: 3,
-    size: "l",
-  },
-  image: {
-    src: "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
-    alt: "The National Archives office",
-    width: 1996,
-    height: 1331,
-  },
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "l",
+  imageSrc:
+    "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
+  imageAlt: "The National Archives office",
+  imageWidth: 1996,
+  imageHeight: 1331,
   label: "New",
   body: "<p>Card body</p>",
   actions: [
@@ -161,18 +165,15 @@ Horizontal.args = {
 
 export const HorizontalBoxed = Template.bind({});
 HorizontalBoxed.args = {
-  heading: {
-    supertitle: "Card supertitle",
-    title: "Card title",
-    level: 3,
-    size: "l",
-  },
-  image: {
-    src: "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
-    alt: "The National Archives office",
-    width: 1996,
-    height: 1331,
-  },
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "l",
+  imageSrc:
+    "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
+  imageAlt: "The National Archives office",
+  imageWidth: 1996,
+  imageHeight: 1331,
   label: "New",
   body: "<p>Card body</p>",
   actions: [
@@ -191,18 +192,15 @@ HorizontalBoxed.args = {
 
 export const HorizontalAccent = Template.bind({});
 HorizontalAccent.args = {
-  heading: {
-    supertitle: "Card supertitle",
-    title: "Card title",
-    level: 3,
-    size: "l",
-  },
-  image: {
-    src: "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
-    alt: "The National Archives office",
-    width: 1996,
-    height: 1331,
-  },
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "l",
+  imageSrc:
+    "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
+  imageAlt: "The National Archives office",
+  imageWidth: 1996,
+  imageHeight: 1331,
   label: "New",
   body: "<p>Card body</p>",
   actions: [
@@ -219,10 +217,49 @@ HorizontalAccent.args = {
   classes: "tna-card--demo",
 };
 
+export const Sources = Template.bind({});
+Sources.args = {
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "m",
+  imageSrc: "https://www.gstatic.com/webp/gallery/2.jpg",
+  imageAlt: "A man in a canoe paddling through white water",
+  imageWidth: 550,
+  imageHeight: 404,
+  imageSources: [
+    {
+      imageSrc: "https://www.gstatic.com/webp/gallery/2.webp",
+      type: "image/webp",
+    },
+  ],
+  body: "<p>Card body</p>",
+  htmlElement: "article",
+  classes: "tna-card--demo",
+};
+
+export const WithoutImage = Template.bind({});
+WithoutImage.args = {
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "m",
+  body: "<p>Card body</p>",
+  htmlElement: "article",
+  classes: "tna-card--demo",
+};
+
 const GridTemplate = ({
-  heading,
+  title,
+  supertitle,
+  headingLevel,
+  headingSize,
   href,
-  image,
+  imageSrc,
+  imageAlt,
+  imageWidth,
+  imageHeight,
+  imageSources,
   label,
   body,
   text,
@@ -238,9 +275,16 @@ const GridTemplate = ({
         `<div class="tna-column tna-column--width-1-3 tna-column--width-1-2-small tna-column--full-tiny tna-!--margin-bottom-m">
           ${Card({
             params: {
-              heading,
+              title,
+              supertitle,
+              headingLevel,
+              headingSize,
               href,
-              image,
+              imageSrc,
+              imageAlt,
+              imageWidth,
+              imageHeight,
+              imageSources,
               label,
               body,
               text,
@@ -258,28 +302,18 @@ const GridTemplate = ({
 
 export const Grid = GridTemplate.bind({});
 Grid.args = {
-  heading: {
-    supertitle: "Card supertitle",
-    title: "Card title",
-    level: 3,
-    size: "m",
-  },
+  supertitle: "Card supertitle",
+  title: "Card title",
+  headingLevel: 3,
+  headingSize: "m",
   href: "#",
-  image: {
-    src: "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
-    alt: "The National Archives office",
-    width: 1996,
-    height: 1331,
-  },
+  imageSrc:
+    "https://www.nationalarchives.gov.uk/wp-content/uploads/sites/24/2023/07/tna-building-compress.jpg",
+  imageAlt: "The National Archives office",
+  imageWidth: 1996,
+  imageHeight: 1331,
   label: "New",
   body: "<p>Card body</p>",
-  // actions: [
-  //   {
-  //     text: "Card action",
-  //     href: "#",
-  //     title: "Go and do the action",
-  //   },
-  // ],
   htmlElement: "article",
   classes: "tna-card--demo",
 };